+++ engine/solver.py
@@ -36,7 +36,7 @@ class Solver:
         # allow up to 5 chances for convergence - above 20 doesnt lead to many additional convergences
         guesses_allowed = 5
         for i in range(guesses_allowed):
-            results  = fsolve(self.__DOF6_motion_residuals, self.__initial_guess, full_output = True)
+            results = fsolve(self.__DOF6_motion_residuals, self.__initial_guess, full_output = True)
             if results[2] == 1:
                 if i != 0:
                     print("Solution converged after changing initial guess")
@@ -45,7 +45,7 @@ class Solver:
                     # NOTE: Solution converged on first guess!
                 return copy(self.vehicle.logger.return_log())
             elif results[2] != 1:
-                if i == (guesses_allowed -1 ):
+                if i == (guesses_allowed - 1):
                     print(f"Solution convergence not found after {guesses_allowed} guesses for state: {input_state.body_slip} {input_state.s_dot} {input_state.steered_angle}")
                     #print(results[1]["fvec"],"\n") # for debugging why the solution didnt converge
                     return {}
